+++ packages/@ionic/cli/src/lib/project/index.ts
@@ -4,7 +4,7 @@ import { resolveValue } from '@ionic/cli-framework/utils/fn';
 import { ERROR_INVALID_PACKAGE_JSON, compileNodeModulesPaths, isValidPackageName, readPackageJsonFile } from '@ionic/cli-framework/utils/node';
 import { ensureDir, findBaseDirectory, readFile, writeFile, writeJson } from '@ionic/utils-fs';
 import { TTY_WIDTH, prettyPath, wordWrap } from '@ionic/utils-terminal';
-import * as Debug from 'debug';
+import { debug as Debug } from 'debug';
 import * as lodash from 'lodash';
 import * as path from 'path';